+++ lib/pio/open_flow10.rb
@@ -27,8 +27,8 @@ require 'pio/open_flow10/stats_request'
 require 'pio/open_flow10/table_stats/request'
 
 # Actions
-require 'pio/open_flow10/nicira_resubmit'
-require 'pio/open_flow10/nicira_resubmit_table'
+require 'pio/open_flow/nicira_resubmit'
+require 'pio/open_flow/nicira_resubmit_table'
 require 'pio/open_flow10/send_out_port'
 require 'pio/open_flow10/set_destination_mac_address'
 require 'pio/open_flow10/set_source_mac_address'
