+++ lib/pio/open_flow13.rb
@@ -13,6 +13,8 @@ require 'pio/open_flow13/packet_out'
 require 'pio/open_flow13/stats_request'
 
 # Actions
+require 'pio/open_flow/nicira_resubmit'
+require 'pio/open_flow/nicira_resubmit_table'
 require 'pio/open_flow13/nicira_reg_move'
 require 'pio/open_flow13/send_out_port'
 require 'pio/open_flow13/set_arp_operation'
